requirements.txt
alembic==1.7.7
bcrypt==3.2.2
blinker==1.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.1
colorama==0.4.4
cssbeautifier==1.14.7
dnspython==2.1.0
EditorConfig==0.12.3
email-validator==1.1.3
Faker==18.9.0
Flask==2.0.1
Flask-Admin==1.6.0
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsbeautifier==1.14.7
Mako==1.2.0
MarkupSafe==2.0.1
mysqlclient==2.1.0
packaging==23.1
pathspec==0.11.1
Pillow
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.1
PyYAML==6.0
regex==2023.5.5
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.23
text-unidecode==1.3
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.9
Werkzeug==2.0.1
WTForms==2.3.3